requirements.txt
langchain
langchain_community
pypdf
langchain-google-genai
python-dotenv
langchain-ollama
faiss-cpu
sentence-transformers
gradio
langchain_groq
streamlit